requirements.txt
asgiref==3.2.7
autopep8==1.4.4
Django==2.2.3
django-crispy-forms==1.7.2
djangorestframework==3.11.0
pycodestyle==2.5.0
pytz==2019.1
sqlparse==0.3.0
